Update final.html
-Small Fix To final.html 2nd title text

View previous change log for major details:
-------------------
Major Update officially introducing the limited QWERTY keyboard support for the visuals. This gives the user an option to play around before plugging in and setting up their midi controller. I have also made improvements to the usability for people with one MIDI controller as a bug fix within the midi file was only active if you changed the MIDI controllers you were going to use. This is what the fix MIDI button is for.
-There is also a range feature which calculates the range of the notes being played and scatters particles across that range.
-There is also a a fix the the particles that would spawn at the lowest possible note whenever any new note was pressed no matter the value
-There is also much more instructions provided on screen to help users get started
Next I have plans to:
-make the range visible under the MIDI NOTE NUMBERS and VELOCITIES text above the visuals
-create something translates MIDI note numbers onscreen into note values for the user